pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frontier-ai-radar"
version = "0.1.0"
description = "Daily multi-agent intelligence system for tracking frontier AI developments"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "prefect>=3.0.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "playwright>=1.48.0",
    "trafilatura>=1.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "feedparser>=6.0.0",
    "pdfminer.six>=20221105",
    "aiohttp>=3.10.0",
    "python-multipart>=0.0.12",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]
